fix some circular dependencies
