+++ .madgerc
@@ -1,0 +1,9 @@
+{
+    "fileExtensions" : ["ts"],
+	"detectiveOptions": {
+		"ts": {
+			"skipTypeImports": true,
+			"mixedImports": true
+		}
+	}
+}
