+++ packages/core/src/header-builder/authorization-header.ts
@@ -13,7 +13,7 @@ import {
   sanitizeDestination
 } from '../scp-cf';
 import { ODataRequestConfig } from '../odata/common/request';
-import { ODataRequest } from '../odata/common/request/odata-request';
+import type { ODataRequest } from '../odata/common/request/odata-request';
 import { getHeader, toSanitizedHeaderObject } from './headers-util';
 
 const logger = createLogger({
