+++ packages/core/src/odata/common/filter/filter-link.ts
@@ -1,9 +1,8 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
-import { Constructable } from '../constructable';
-import { EntityBase, EntityIdentifiable } from '../entity';
+import { Constructable, EntityBase, EntityIdentifiable } from '../entity';
 import { Link } from '../selectable';
-import { Filterable } from './filterable';
+import type { Filterable } from './filterable';
 
 /**
  * Data structure to represent filter on properties of a navigation property (link).
