+++ packages/core/src/odata/common/index.ts
@@ -1,7 +1,6 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
 export * from './batch-response';
-export * from './constructable';
 export * from './edm-types';
 export * from './entity';
 export * from './entity-builder';
