+++ packages/core/src/odata/common/request-builder/request-builder-base.ts
@@ -117,3 +117,4 @@ function noDestinationErrorMessage(
     ? `Could not find a destination with name "${destination.destinationName}"! Unable to execute request.`
     : 'Could not find a destination to execute request against and no destination name has been provided (this should never happen)!';
 }
+
