+++ packages/core/src/odata/common/request-builder/request-builder.ts
@@ -1,7 +1,7 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
-import { Constructable } from '../constructable';
-import { EntityBase, EntityIdentifiable } from '../entity';
+import { EntityBase, EntityIdentifiable, Constructable } from '../entity';
+
 
 type EntityBasedRequestBuilder<
   EntityCT extends Constructable<EntityBase>
