+++ packages/core/src/odata/common/request/odata-delete-request-config.ts
@@ -1,8 +1,7 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
 import { MapType } from '@sap-cloud-sdk/util';
-import { Constructable } from '../constructable';
-import { EntityBase } from '../entity';
+import { Constructable, EntityBase } from '../entity';
 import { FieldType } from '../selectable';
 import { ODataUri } from '../uri-conversion';
 import { ODataRequestConfig } from './odata-request-config';
