+++ packages/core/src/odata/common/request/odata-get-by-key-request-config.ts
@@ -1,8 +1,7 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
 import { MapType } from '@sap-cloud-sdk/util';
-import { Constructable } from '../constructable';
-import { EntityBase } from '../entity';
+import { EntityBase, Constructable } from '../entity';
 import { FieldType, Selectable } from '../selectable';
 import { Expandable } from '../expandable';
 import { ODataUri } from '../uri-conversion';
