+++ packages/core/src/odata/common/selectable/all-fields.ts
@@ -1,7 +1,6 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
-import { Constructable } from '../constructable';
-import { EntityBase } from '../entity';
+import { EntityBase, Constructable } from '../entity';
 
 export class AllFields<EntityT extends EntityBase> {
   readonly selectable: true;
