+++ packages/core/src/odata/common/selectable/complex-type-field.ts
@@ -1,9 +1,8 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
 import { ODataVersion } from '@sap-cloud-sdk/util';
-import { Constructable } from '../constructable';
 import { EdmTypeShared } from '../edm-types';
-import { EntityBase } from '../entity';
+import { Constructable, EntityBase } from '../entity';
 import { Field } from './field';
 
 /**
