+++ packages/core/src/odata/common/selectable/field.ts
@@ -2,8 +2,8 @@
 
 import BigNumber from 'bignumber.js';
 import { Moment } from 'moment';
-import { Constructable } from '../constructable';
-import { EntityBase, EntityIdentifiable } from '../entity';
+import { EntityBase, EntityIdentifiable, Constructable } from '../entity';
+
 import { Time } from '../time';
 
 /**
