+++ packages/core/src/odata/common/selectable/string-field.ts
@@ -1,9 +1,8 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 /* eslint-disable max-classes-per-file */
 
-import { Constructable } from '../constructable';
 import { EdmTypeShared } from '../edm-types';
-import { EntityBase, ODataVersionOf } from '../entity';
+import { EntityBase, ODataVersionOf, Constructable } from '../entity';
 import {
   ComplexTypeField,
   ConstructorOrField,
