+++ packages/core/src/odata/v2/odata-batch-change-set.ts
@@ -6,12 +6,12 @@ import {
   content_transfer_encoding_line
 } from '../common';
 import { Entity } from './entity';
-import { getRequestLine } from './odata-batch-request-builder';
 import {
   CreateRequestBuilder,
   DeleteRequestBuilder,
   UpdateRequestBuilder
 } from './request-builder';
+import { getRequestLine } from '../common/request/odata-batch-request-util';
 
 const batch_content_type_prefix =
   'Content-Type: multipart/mixed; boundary=changeset_';
