+++ packages/core/src/odata/v4/odata-batch-request-builder.ts
@@ -11,7 +11,6 @@ import {
 } from '../../scp-cf';
 import {
   Constructable,
-  http_version,
   BatchResponse,
   WriteResponse,
   WriteResponses,
@@ -20,7 +19,6 @@ import {
 } from '../common';
 import { MethodRequestBuilderBase } from '../common/request-builder/request-builder-base';
 import { ODataBatchRequestConfig } from '../common/request/odata-batch-request-config';
-import { ODataRequestConfig } from '../common/request/odata-request-config';
 import { Entity } from './entity';
 import { deserializeEntity } from './entity-deserializer';
 import {
@@ -364,16 +362,6 @@ function getEndBatchWithLineBreak(batchId: string) {
   return `--batch_${batchId}--\n`;
 }
 
-function getMethod(
-  requestBuilder: MethodRequestBuilderBase<ODataRequestConfig>
-) {
-  return requestBuilder.requestConfig.method;
-}
-
-function getUrl(requestBuilder: MethodRequestBuilderBase<ODataRequestConfig>) {
-  return `/${requestBuilder.relativeUrl()}`;
-}
-
 function isChangeSet(response: string): boolean {
   return !!response.match(/Content-Type: multipart\/mixed;/);
 }
@@ -467,16 +455,4 @@ function buildRetrieveOrErrorResponse(
   return { httpCode, body: parsedBody, isSuccess: () => false };
 }
 
-/**
- * Generate the request line, containing method, url and http version from the request builder, e.g.:
- * GET /sap/opu/odata/sap/API_BUSINESS_PARTNER/A_BusinessPartnerAddress?$format=json&$top=1 HTTP/1.1
- * @param requestBuilder - Reqeust builder holds the request information.
- * @returns the generated request line.
- */
-export function getRequestLine(
-  requestBuilder: MethodRequestBuilderBase<ODataRequestConfig>
-): string {
-  return `${getMethod(requestBuilder).toUpperCase()} ${getUrl(
-    requestBuilder
-  )} ${http_version}`;
-}
+
