+++ test-packages/test-services/srv/test-service-legacy/builder.ts
@@ -1,0 +1,5 @@
+import { TestEntityMultiLinkRequestBuilder } from '../v2/test-service';
+
+export function getFoo():TestEntityMultiLinkRequestBuilder{
+  return new TestEntityMultiLinkRequestBuilder();
+}
